ADDED: DDT_SUB_ELEMENTS filter gets the element's key and binder
filter are called at different times while generating the dump

Data::Dump::Tree cycle is:

when the element is to be displayed, DDT calls the element which return
a description of itself then
        * filters DDT_HEADER are called

when sub elements of an element are to be displayed, DDT calls the element
which returns a list of the sub elements in the form ('name' 'binder' 'sub_element')
        * filters DDT_SUB_ELEMENTS are called

when DDT has rendered the element and will get to the next element
        * filters DDT_FOOTER are called

An element doesn't know under which name it will be displayed has the name
is decided by its parent. The same element could be present multiple times
in the data structure, each time under a different parent which may give it
a different name

in this commit the 'name' given by the parent is passed to filters both filter
DDT_HEADER and DDT_SUB_ELEMENTS.

This was added to better the output of grammar parsing. Grammars return Match
object that may themselves contain sub Match objects; as all are Match objects,
it is difficult to filter them without knowing what Match object matched, we
get the information from the 'name' given by the parent. There may be better
ways but this is simple and efficient without having any impact outside the
dumper. It does need a deeper knowledge of the dumped data.

You can see an example of a filter using the 'name' in example/JSON_parsed.pl

With two simple filters the output length is halved without losing any information

doc